+++ NEWS.md
@@ -6,6 +6,7 @@
 * Added `remove_extraneous` to exported functions, for removing redundant dependencies or determinant attributes in a `functional_dependency` object. This is used as a step in `synthesise`, but is useful on its own.
 * Added support for data with list columns. `duplicate`, `unique` etc. consider data frame rows different if they only differ by `NA` value classes in list columns. Previously, `discover` considered them to be equal, so a data frame with list columns could violate its own schema. This is now fixed, in exchange for a small performance hit when there are list columns.
 * `rejoin` no longer sorts as it merges relations, since there might be list columns in the key, which aren't sortable.
+* Added nesting to `gv` and `d2` methods for classes containing data. This allows giving more detail when a data column is a list where elements have a common class or length.
 
 # autodb 3.2.4